Put things in order in folders: put the classes in packages by functional, rename "dummy" to "data".
