+++ app/src/main/java/com/example/myproject/project/list/NoteAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.myproject.project
+package com.example.myproject.project.list
 
 import android.view.LayoutInflater
 import android.view.View
@@ -9,30 +9,33 @@ import android.widget.TextView
 import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.currentnote.R
+import com.example.myproject.project.note.Note
 import com.example.myproject.project.util.DateFormatter
+import com.example.myproject.project.wallpaper.Wallpaper
 import kotlin.collections.ArrayList
 
 
-class HiddenNoteAdapter(private val showDetailListener: ShowDetailListener) :
-        RecyclerView.Adapter<HiddenNoteAdapter.HiddenNoteHolder>() {
+class NoteAdapter(_showDetail: ShowDetail) : RecyclerView.Adapter<NoteAdapter.NoteHolder>() {
     private val noteList = ArrayList<Note>()
+    private val showDetail = _showDetail
     private var isVisible = false
     private var checked = ArrayList<Boolean>()
     private val checkList = ArrayList<CheckBox>()
 
-    interface ShowDetailListener {
+    interface ShowDetail {
         fun onClickElement(note: Note? = null, position: Int = 0)
 
         fun onLongClickElement()
     }
 
-    class HiddenNoteHolder(item: View) : RecyclerView.ViewHolder(item) {
+    class NoteHolder(item: View) : RecyclerView.ViewHolder(item) {
+        private val cardView: CardView = item.findViewById(R.id.cardView)
         private val tvTitle: TextView = item.findViewById(R.id.tvTitleNote)
         private val tvContent: TextView = item.findViewById(R.id.tvContentNote)
         private val tvTime: TextView = item.findViewById(R.id.tvTime)
         val checkBox: CheckBox = item.findViewById(R.id.checkRec)
         private val pin: ImageView = item.findViewById(R.id.imTop)
-        private val cardView: CardView = item.findViewById(R.id.cardView)
+
 
         fun initNote(note: Note) {
             tvTitle.text = note.title
@@ -41,23 +44,25 @@ class HiddenNoteAdapter(private val showDetailListener: ShowDetailListener) :
             if (note.isTop) {
                 pin.visibility = View.VISIBLE
             } else pin.visibility = View.GONE
+
             if (note.wallpaperName != null) {
                 val wallpaper: Wallpaper = Wallpaper.valueOf(note.wallpaperName)
-                tvTitle.setTextColor(wallpaper.textColor)
-                tvContent.setTextColor(wallpaper.textColor)
-                cardView.setBackgroundResource(wallpaper.primaryBackground)
-                tvTime.setTextColor(wallpaper.textColor)
+                    tvTitle.setTextColor(wallpaper.textColor)
+                    tvContent.setTextColor(wallpaper.textColor)
+                    cardView.setBackgroundResource(wallpaper.primaryBackground)
+                    tvTime.setTextColor(wallpaper.textColor)
             }
         }
     }
 
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HiddenNoteHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NoteHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.my_note_item, parent, false)
-        return HiddenNoteHolder(view)
+        return NoteHolder(view)
     }
 
-    override fun onBindViewHolder(holder: HiddenNoteHolder, position: Int) {
+    override fun onBindViewHolder(holder: NoteHolder, position: Int) {
+
         if (isVisible) {
             checkList.add(holder.checkBox)
             holder.checkBox.visibility = View.VISIBLE
@@ -68,7 +73,7 @@ class HiddenNoteAdapter(private val showDetailListener: ShowDetailListener) :
 
         holder.checkBox.setOnClickListener {
             checked[position] = !checked[position]
-            showDetailListener.onClickElement()
+            showDetail.onClickElement()
         }
 
         holder.initNote(noteList[position])
@@ -84,16 +89,16 @@ class HiddenNoteAdapter(private val showDetailListener: ShowDetailListener) :
             val isTop = noteList[holder.absoluteAdapterPosition].isTop
             val wallpaperName = noteList[holder.absoluteAdapterPosition].wallpaperName
             val note = Note(title, content, id, time, isTop, wallpaperName)
-            showDetailListener.onClickElement(note, elementPosition)
+            showDetail.onClickElement(note, elementPosition)
             if (isVisible) {
                 checked[position] = !checked[position]
                 holder.checkBox.isChecked = checked[position]
-                showDetailListener.onClickElement()
+                showDetail.onClickElement()
             }
         }
 
         holder.itemView.setOnLongClickListener {
-            showDetailListener.onLongClickElement()
+            showDetail.onLongClickElement()
             true
         }
 
@@ -133,4 +138,5 @@ class HiddenNoteAdapter(private val showDetailListener: ShowDetailListener) :
         noteList.addAll(listItems)
         notifyDataSetChanged()
     }
+
 }
