+++ app/src/main/java/com/example/myproject/project/trash/TrashFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myproject.project
+package com.example.myproject.project.trash
 
 import android.annotation.SuppressLint
 import android.app.AlertDialog
@@ -13,7 +13,8 @@ import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.example.myproject.project.util.OnBackPressedListener
 import com.example.currentnote.R
 import com.example.currentnote.databinding.FragmentTrashBinding
-import com.example.myproject.dummy.dataBase.DbManager
+import com.example.myproject.data.db.DbManager
+import com.example.myproject.project.note.Note
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
