+++ app/src/main/java/com/example/myproject/project/wallpaper/Wallpaper.kt
@@ -1,4 +1,4 @@
-package com.example.myproject.project
+package com.example.myproject.project.wallpaper
 
 import com.example.currentnote.R
