+++ app/src/main/res/layout/fragment_hidden_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.myproject.project.HiddenNotesFragment">
+    tools:context="com.example.myproject.project.hidden.HiddenNotesFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout"
