+++ app/src/main/res/layout/fragment_trash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.myproject.project.TrashFragment">
+    tools:context="com.example.myproject.project.trash.TrashFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
